Crate problem126

Problem 126 (Cuboid layers)

Solution detail

It's slightly fiddly to calculate the number of cubes in each layer, but it can be done! If the side lengths of the cuboid are a, b and c then the number of cubes in a single layer is:

2(ab + bc + ca) + 4(n-1)(a + b + c - 1) + 4(n-1)²

Then, brute force is our best friend. Assuming that the answer is below a particular limit, then just iterate over all cuboids, and over all layers, storing the number of times each number of cubes has been seen, for each number of cubes up to the limit. If any number of cubes has been used the right number of times, then that's the answer. Otherwise, increase the limit a little (e.g. double it) and start again.

Functions

answer
Solve the problem, returning the answer as a String

layer_size_counts
Calculate the number of times each layer size is used, up to a limit.

main

smallest_size_with_n_occurrences
Find the first layer size which occurs exactly n times, among all cuboids.

solve
Find the smallest layer size which appears exactly the given number of times.